requirements.txt
glob3==0.0.1
numpy==1.16.4
opencv-python==4.1.1.26
Pillow==5.3.0
torch==1.2.0+cpu
torchvision==0.4.0+cpu
tqdm==4.43.0
